feat: exception mapper for domains exception

+++ src/main/java/application/handlers/DeleteUserHandler.java
@@ -2,6 +2,7 @@ package application.handlers;
 
 import application.commands.DeleteUserCommand;
 import application.commands.FindUserCommand;
+import application.domain.exceptions.UserNotFoundException;
 import application.domain.models.Cpf;
 import application.domain.models.User;
 import application.ports.inbound.DeleteUser;
@@ -23,6 +24,8 @@ public class DeleteUserHandler implements DeleteUser {
     @Override
     public void execute(DeleteUserCommand command) {
         Cpf cpf = command.cpf;
+        Optional<User> user = this.repository.findBy(cpf);
+        if (user.isEmpty()) throw new UserNotFoundException(cpf);
         repository.deleteBy(cpf);
     }
 }
